initial quest id generation
