+++ package.json
@@ -7,7 +7,8 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "prepare": "husky install"
+    "prepare": "husky install",
+    "postinstall": "npx prisma generate"
   },
   "dependencies": {
     "@fontsource/inter": "^4.5.7",
@@ -30,6 +31,7 @@
     "@types/react": "18.0.5",
     "@types/react-beautiful-dnd": "^13.1.2",
     "@types/react-dom": "18.0.0",
+    "@types/uuid": "^8.3.4",
     "autoprefixer": "^10.4.4",
     "eslint": "8.13.0",
     "eslint-config-next": "12.1.5",
@@ -39,6 +41,7 @@
     "postcss": "^8.4.12",
     "prettier": "^2.6.2",
     "prisma": "^3.12.0",
+    "prisma-yup-generator": "^0.0.0-rc.2",
     "tailwindcss": "^3.0.24",
     "typescript": "4.6.3"
   }
