+++ prisma/schema.prisma
@@ -45,6 +45,7 @@ model User {
   image         String?
   accounts      Account[]
   sessions      Session[]
+  Quests        Quest[]
 }
 
 model VerificationToken {
@@ -53,4 +54,17 @@ model VerificationToken {
   expires    DateTime
 
   @@unique([identifier, token])
+}
+
+model Quest {
+  id          String  @id @default(uuid())
+  name        String?
+  description String?
+  draft       Boolean @default(true)
+  user        User    @relation(fields: [userId], references: [id])
+  userId      String
+}
+
+generator yup {
+  provider = "prisma-yup-generator"
 }
