+++ types/index.ts
@@ -1,0 +1,3 @@
+export type APIError = {
+  error: any
+}
